requirements.txt
transformers
torch
Pillow
fastapi
uvicorn
